refactor: changing promise negotiation
