--- channel-program/.env.example
@@ -1,2 +1,0 @@
-PUBLIC_URL=https://[your_given_asset_host_path].platform-os.com/instances/[instance_id]/assets/web-app/
-REACT_APP_POS_PROXY=https://wispy-sun-870.staging.oregon.platform-os.com/
